pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "context-reliability-testing"
version = "0.1.0"
description = "A/B test agent context files (AGENTS.md, .cursorrules) against real tasks on your repo."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "typer>=0.12",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]
assertions = ["pytest>=8.0"]

[project.scripts]
crt = "context_reliability_testing.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/context_reliability_testing"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
